% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{Apply a function on each column of a data frame given a condition}
\usage{
df_apply(.data, .f, .condition = is.numeric, .else = identity, ...)
}
\arguments{
\item{.data}{A data frame}

\item{.f}{A function to apply on the data frame}

\item{.condition}{Default value "is.numeric", but can be another data type}

\item{.else}{Default value "identity" (skips over a column)}

\item{...}{Any arguments from .f to pass}
}
\value{
A tibble
}
\description{
Apply a function on each column of a data frame given a condition
}
